AI/AI01/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
numpy==1.21.2
Pillow==8.3.1
torch==1.9.0+cpu
torchvision==0.10.0+cpu
torchaudio==0.9.0
tqdm==4.62.2
